requirements.txt
fastmcp
requests
torch
datasets
peft
transformers
trl
bitsandbytes
PyYAML
cryptography